pyproject.toml
[project]
name = "private-nba-analytics"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "altair>=5.0.0",
    "db-dtypes>=1.5.0",
    "google-cloud-bigquery>=3.25.0",
    "ipykernel>=7.2.0",
    "nba-api>=1.11.4",
    "pandas<3.0.0",
    "pandas-gbq>=0.13.2",
    "pyarrow>=11.0.0",
    "python-dotenv>=1.0.0",
    "requests>=2.32.5",
    "streamlit>=1.30.0",
]
